pyproject.toml
[tool.poetry]
name = "channel-dublicator"
version = "0.1.0"
description = ""
authors = ["wannawhat <wannawhere4@gmail.com>"]
readme = "README.md"
packages = [{include = "channel_dublicator"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
pydantic = "^1.10.9"
pyrogram = "^2.0.106"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

